fixed symlinking of dirs

+++ helper/remove.sh
@@ -3,11 +3,25 @@
 . ./helper/cmd.sh
 
 remove() {
-  name="Remove $(basename "$1")"
-  eval_cmd "$name" "rm -rf $1"
+  if [ -d "$1" ]; then
+    name="Remove directory $(basename "$1")"
+    eval_cmd "$name" "rm -rf $1"
+  fi
+
+  if [ -d "$1" ]; then
+    name="Remove file $(basename "$1")"
+    eval_cmd "$name" "rm -rf $1"
+  fi
 }
 
 sudo_remove() {
-  name="Remove $(basename "$1")"
-  eval_cmd "$name" "sudo rm -rf $1"
+  if [ -d "$1" ]; then
+    name="Remove directory $(basename "$1")"
+    eval_cmd "$name" "sudo rm -rf $1"
+  fi
+
+  if [ -d "$1" ]; then
+    name="Remove file $(basename "$1")"
+    eval_cmd "$name" "sudo rm -rf $1"
+  fi
 }
